# docs/api/order-service-openapi.yaml
openapi: 3.0.1
info:
  title: Order Service API
  version: 1.0.0
  description: API reactiva para gestión de órdenes de e-commerce.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /orders:
    post:
      summary: Crear nueva orden
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Orden creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    get:
      summary: Listar órdenes de un cliente
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de órdenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
  /orders/{id}:
    get:
      summary: Obtener orden por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Orden no encontrada
  /orders/{id}/cancel:
    patch:
      summary: Cancelar orden (PENDING o CONFIRMED)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Orden cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Orden no encontrada
        '409':
          description: Estado inválido para cancelación
  /orders/{id}/events:
    get:
      summary: Historial de eventos de la orden
      parameters:
        - name: id
          in: path
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderEventResponse'
components:
  schemas:
    CreateOrderRequest:
      type: object
      properties:
        customerId:
          type: string
        totalAmount:
          type: number
          format: double
      required:
        - customerId
        - totalAmount
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderEventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        type:
          type: string
        occurredAt:
          type: string
          format: date-time
        payload:
          type: object